change where_op to where_v2_op, where_index_op to where_op.
